@@ -2,7 +2,7 @@ package com.ttt.capstone.request;
 
 import lombok.*;
 
-import javax.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotBlank;
 
 @Getter
 @ToString
